Independently vendor stb libraries
This is for one reason, and one reason alone:

stb_truetype doesn't include stb_rect_pack, therefore not using the new
rect packing API, thusly making it incompatible. So we instead include
it in the build static libraries, but not it's implementation as to not
conflict with stb_rect_pack itself.